% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3sink.R
\name{sink_s3}
\alias{sink_s3}
\alias{unsink_s3}
\title{Capture arguments to S3 generics}
\usage{
sink_s3(name, pkgname = NULL)

unsink_s3(name)
}
\arguments{
\item{name}{character string. Name of an S3 generic}

\item{pkgname}{character string. Name of a package which registers the
generic used in \code{name}. (optional)}
}
\value{
\code{sink_s3} returns invisibly \code{NULL}

\code{unsink_s3} returns invisibly a list of lists of arguments
}
\description{
Capture arguments to S3 generics
}
\details{
...
}
\examples{
sink_s3("print.ggplot", "ggplot2")
library(ggplot2)
ggplot(mtcars) + geom_point(aes(mpg, disp))
out <- unsink_s3("print.ggplot")

str(out)
}
